Implement Focus Package
