+++ src/main/java/Focus/Focus.java
@@ -1,3 +1,5 @@
+package Focus;
+
 import java.io.IOException;
 
 public class Focus {
