+++ src/main/java/Focus/Task.java
@@ -1,3 +1,5 @@
+package Focus;
+
 /**
  * Represents a task
  */
