+++ src/main/java/Focus/TodoCommand.java
@@ -1,3 +1,5 @@
+package Focus;
+
 public class TodoCommand extends FocusCommand {
 
     private final String description;
